Large refactor of test framework
Makes many changes to test infrastructure:
* Split test_util into separate files for maintainability
* Move new files into tests/framework/util
* Move FolderManager into separate file from test_environment
* Rename COMMON_UNIX_PLATFORMS to TESTING_COMMON_UNIX_PLATFORMS
* Rewrote unix path redirection logic, relies on FileSystemManager
  * Relative paths are actually relative now
* Split unsecured folder location into unsecured driver/implicit layer/explicit layer/settings
* Fix bugs in tests that are attributable to old path redirection logic.
* Use normalized paths in test framework - ie normalize paths so they can be compared easily
* Make running with elevated privileges a framework level setting
* Split tests that change elevated privileges into separate tests
* Framework can set home & xdg env-vars at startup
* Platform specific folders are only defined on those platforms
* PlatformShim::add_manifest is now windows only, reflecting that only windows implemented it
