+++ scripts/generators/dispatch_table_helper_generator.py
@@ -60,7 +60,7 @@ class DispatchTableHelperGenerator(BaseGenerator):
 #include <vulkan/vulkan.h>
 #include <vulkan/vk_layer.h>
 #include <string.h>
-#include "vk_layer_dispatch_table.h"
+#include "loader/generated/vk_layer_dispatch_table.h"
 
 ''')
