+++ tests/framework/CMakeLists.txt
@@ -15,38 +15,15 @@
 # limitations under the License.
 # ~~~
 
-add_library(testing_framework_util STATIC test_util.cpp)
-target_link_libraries(testing_framework_util PUBLIC loader_common_options Vulkan::Headers)
+#configure the framework_config.h.in file - used to locate all the binaries generated so that it can be used in the tests
+#setup framework_config_temp.h.in in the current binary directory
+configure_file("${CMAKE_CURRENT_SOURCE_DIR}/framework_config.h.in" "${CMAKE_CURRENT_BINARY_DIR}/framework_config_temp.h.in")
 
-if(UNIX OR APPLE)
-    target_link_libraries(testing_framework_util PUBLIC ${CMAKE_DL_LIBS})
-endif()
+# setup framework_config_$<CONFIG> using framework_config_temp.h.in as a source
+file(GENERATE OUTPUT "${CMAKE_CURRENT_BINARY_DIR}/framework_config_$<CONFIG>.h" INPUT "${CMAKE_CURRENT_BINARY_DIR}/framework_config_temp.h.in")
 
-if(UNIX)
-    target_compile_options(testing_framework_util PUBLIC -fPIC)
-endif()
-# Gives access to all headers in the framework folder, in the framework binary, and in the whole project (mainly for loader/generated)
-target_include_directories(testing_framework_util PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}" "${CMAKE_CURRENT_BINARY_DIR}" "${PROJECT_SOURCE_DIR}" "${PROJECT_SOURCE_DIR}/loader/generated")
-
-if (UNIX)
-    if (LOADER_ENABLE_ADDRESS_SANITIZER)
-        target_compile_options(testing_framework_util PUBLIC -fsanitize=address,undefined)
-        target_link_options(testing_framework_util PUBLIC -fsanitize=address,undefined)
-    endif()
-    if (LOADER_ENABLE_THREAD_SANITIZER)
-        target_compile_options(testing_framework_util PUBLIC -fsanitize=thread)
-        target_link_options(testing_framework_util PUBLIC -fsanitize=thread)
-        target_compile_options(gtest PUBLIC -fsanitize=thread)
-        target_link_options(gtest PUBLIC -fsanitize=thread)
-    endif()
-endif()
-
-if (MSVC)
-    # silence hidden class member warnings in test framework
-    target_compile_options(testing_framework_util PUBLIC /wd4458)
-    # Make sure exception handling is enabled for the test framework
-    target_compile_options(testing_framework_util PUBLIC /EHsc)
-endif()
+# Create a variable to hold the path to the correct header file, used as a compiler definition in the utils target
+set(FRAMEWORK_CONFIG_HEADER_PATH "${CMAKE_CURRENT_BINARY_DIR}/framework_config_$<CONFIG>.h")
 
 function(AddSharedLibrary LIBRARY_NAME)
     set(singleValueArgs DEF_FILE)
@@ -63,26 +40,14 @@ function(AddSharedLibrary LIBRARY_NAME)
     endif()
 endfunction()
 
+add_subdirectory(util)
 add_subdirectory(data)
 add_subdirectory(shim)
 add_subdirectory(icd)
 add_subdirectory(layer)
 
-#configure the framework_config.h.in file - used to locate all the binaries generated so that it can be used in the tests
-#setup framework_config_temp.h.in in the current binary directory
-configure_file("${CMAKE_CURRENT_SOURCE_DIR}/framework_config.h.in" "${CMAKE_CURRENT_BINARY_DIR}/framework_config_temp.h.in")
-
-# setup framework_config_$<CONFIG> using framework_config_temp.h.in as a source
-file(GENERATE OUTPUT "${CMAKE_CURRENT_BINARY_DIR}/framework_config_$<CONFIG>.h" INPUT "${CMAKE_CURRENT_BINARY_DIR}/framework_config_temp.h.in")
-
-# Add a compiler definition for the path to framework_config.h with the correct config
-target_compile_definitions(testing_framework_util PUBLIC FRAMEWORK_CONFIG_HEADER="framework_config_$<CONFIG>.h")
-
 add_library(testing_dependencies STATIC test_environment.cpp test_environment.h)
 target_link_libraries(testing_dependencies
     PUBLIC gtest Vulkan::Headers testing_framework_util shim-library)
 target_include_directories(testing_dependencies PUBLIC ${CMAKE_CURRENT_BINARY_DIR})
-if (APPLE_STATIC_LOADER)
-    target_compile_definitions(testing_dependencies PUBLIC "APPLE_STATIC_LOADER=1")
-    target_link_libraries(testing_dependencies PUBLIC vulkan)
-endif()
+
