+++ tests/framework/README.md
@@ -136,8 +136,7 @@ There are many utilities that the test framework and tests have access to. These
 * Environment Variable Wrapper: `EnvVarWrapper` for creating, setting, getting, and removing environment variables in a RAII manner
 * Windows API error handling helpers
 * filesystem abstractions:
-  * `create_folder`/`delete_folder`
-  * `FolderManager`
+  * `Folder`
     * Creates a new folder with the given name at construction time.
     * Allows writing manifests and files (eg, icd or layer binaries)
     * Automatically destroys the folder and all contained files at destruction
