+++ tests/framework/framework_config.h.in
@@ -26,7 +26,7 @@
  */
 
 #pragma once
-#define FRAMEWORK_BUILD_DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}"
+#define TEST_EXECUTION_DIRECTORY "${TEST_EXECUTION_DIRECTORY}"
 
 #define FRAMEWORK_VULKAN_LIBRARY_PATH "$<TARGET_FILE:vulkan>"
