+++ tests/framework/layer/test_layer.h
@@ -27,13 +27,17 @@
 
 #pragma once
 
-#include "test_util.h"
-
 #include <functional>
+#include <filesystem>
 
-#include "layer/layer_util.h"
+#include <vulkan/vulkan_core.h>
+#include <vulkan/vk_layer.h>
 
 #include "loader/generated/vk_layer_dispatch_table.h"
+
+#include "util/functions.h"
+
+#include "layer_util.h"
 
 /*
 Interface Version 0
