+++ tests/framework/util/dispatchable_handle.h
@@ -1,0 +1,61 @@
+/*
+ * Copyright (c) 2025 The Khronos Group Inc.
+ * Copyright (c) 2025 Valve Corporation
+ * Copyright (c) 2025 LunarG, Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and/or associated documentation files (the "Materials"), to
+ * deal in the Materials without restriction, including without limitation the
+ * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+ * sell copies of the Materials, and to permit persons to whom the Materials are
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice(s) and this permission notice shall be included in
+ * all copies or substantial portions of the Materials.
+ *
+ * THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+ *
+ * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+ * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE MATERIALS OR THE
+ * USE OR OTHER DEALINGS IN THE MATERIALS.
+ *
+ */
+
+#pragma once
+
+#include <vulkan/vk_icd.h>
+
+#include "test_defines.h"
+
+template <typename T>
+struct FRAMEWORK_EXPORT DispatchableHandle {
+    DispatchableHandle() {
+        auto ptr_handle = new VK_LOADER_DATA;
+        set_loader_magic_value(ptr_handle);
+        handle = reinterpret_cast<T>(ptr_handle);
+    }
+    ~DispatchableHandle() {
+        if (handle) {
+            delete reinterpret_cast<VK_LOADER_DATA*>(handle);
+        }
+        handle = nullptr;
+    }
+    DispatchableHandle(DispatchableHandle const&) = delete;
+    DispatchableHandle& operator=(DispatchableHandle const&) = delete;
+    DispatchableHandle(DispatchableHandle&& other) noexcept : handle(other.handle) { other.handle = nullptr; }
+    DispatchableHandle& operator=(DispatchableHandle&& other) noexcept {
+        if (handle) {
+            delete reinterpret_cast<VK_LOADER_DATA*>(handle);
+        }
+        handle = other.handle;
+        other.handle = nullptr;
+        return *this;
+    }
+    bool operator==(T base_handle) { return base_handle == handle; }
+    bool operator!=(T base_handle) { return base_handle != handle; }
+
+    T handle = nullptr;
+};
