+++ tests/loader_envvar_tests.cpp
@@ -28,6 +28,8 @@
 
 #include "test_environment.h"
 
+#include "util/wide_char_handling.h"
+
 // Don't support vk_icdNegotiateLoaderICDInterfaceVersion
 // Loader calls vk_icdGetInstanceProcAddr second
 // does not support vk_icdGetInstanceProcAddr
@@ -120,18 +122,25 @@ TEST(EnvVarICDOverrideSetup, TestOnlyDriverEnvVar) {
     phys_dev_count = 5;
     ASSERT_EQ(inst2->vkEnumeratePhysicalDevices(inst2.inst, &phys_dev_count, phys_devs_array.data()), VK_SUCCESS);
     ASSERT_EQ(phys_dev_count, 5U);
+}
 
-    env.debug_log.clear();
+// Test VK_DRIVER_FILES environment variable with elelvated privileges
+TEST(EnvVarICDOverrideSetup, TestOnlyDriverEnvVarRunningWithElevatedPrivileges) {
+    FrameworkEnvironment env{FrameworkSettings{}.set_run_as_if_with_elevated_privleges(true)};
+    env.add_icd(TestICDDetails(TEST_ICD_PATH_EXPORT_NONE).set_discovery_type(ManifestDiscoveryType::env_var));
+    env.get_test_icd(0).add_physical_device("pd0");
 
-    env.platform_shim->set_elevated_privilege(true);
+    for (uint32_t add = 0; add < 2; ++add) {
+        env.add_icd(TestICDDetails(TEST_ICD_PATH_EXPORT_NONE).set_discovery_type(ManifestDiscoveryType::env_var))
+            .add_physical_device("pd" + std::to_string(add) + "0")
+            .add_physical_device("pd" + std::to_string(add) + "1");
+    }
 
-    InstWrapper inst3{env.vulkan_functions};
-    FillDebugUtilsCreateDetails(inst3.create_info, env.debug_log);
-    inst3.CheckCreate(VK_ERROR_INCOMPATIBLE_DRIVER);
+    InstWrapper inst{env.vulkan_functions};
+    FillDebugUtilsCreateDetails(inst.create_info, env.debug_log);
+    inst.CheckCreate(VK_ERROR_INCOMPATIBLE_DRIVER);
 
     EXPECT_TRUE(env.debug_log.find("vkCreateInstance: Found no drivers!"));
-
-    env.platform_shim->set_elevated_privilege(false);
 }
 
 // Test VK_DRIVER_FILES environment variable containing a path to a folder
@@ -157,27 +166,30 @@ TEST(EnvVarICDOverrideSetup, TestOnlyDriverEnvVarInFolder) {
             .add_physical_device("pd" + std::to_string(add) + "1");
     }
 
-    env.debug_log.clear();
-
     InstWrapper inst2{env.vulkan_functions};
-    FillDebugUtilsCreateDetails(inst2.create_info, env.debug_log);
     inst2.CheckCreate();
 
     phys_dev_count = 5;
     ASSERT_EQ(inst2->vkEnumeratePhysicalDevices(inst2.inst, &phys_dev_count, phys_devs_array.data()), VK_SUCCESS);
     ASSERT_EQ(phys_dev_count, 5U);
+}
+// Test VK_DRIVER_FILES environment variable containing a path to a folder  with elevated privileges
+TEST(EnvVarICDOverrideSetup, TestOnlyDriverEnvVarInFolderWithElevatedPrivileges) {
+    FrameworkEnvironment env{FrameworkSettings{}.set_run_as_if_with_elevated_privleges(true)};
+    env.add_icd(TestICDDetails(TEST_ICD_PATH_EXPORT_NONE).set_discovery_type(ManifestDiscoveryType::env_var).set_is_dir(false));
+    env.get_test_icd(0).add_physical_device("pd0");
 
-    env.debug_log.clear();
+    for (uint32_t add = 0; add < 2; ++add) {
+        env.add_icd(TestICDDetails(TEST_ICD_PATH_EXPORT_NONE).set_discovery_type(ManifestDiscoveryType::env_var))
+            .add_physical_device("pd" + std::to_string(add) + "0")
+            .add_physical_device("pd" + std::to_string(add) + "1");
+    }
 
-    env.platform_shim->set_elevated_privilege(true);
-
-    InstWrapper inst3{env.vulkan_functions};
-    FillDebugUtilsCreateDetails(inst3.create_info, env.debug_log);
-    inst3.CheckCreate(VK_ERROR_INCOMPATIBLE_DRIVER);
+    InstWrapper inst{env.vulkan_functions};
+    FillDebugUtilsCreateDetails(inst.create_info, env.debug_log);
+    inst.CheckCreate(VK_ERROR_INCOMPATIBLE_DRIVER);
 
     EXPECT_TRUE(env.debug_log.find("vkCreateInstance: Found no drivers!"));
-
-    env.platform_shim->set_elevated_privilege(false);
 }
 
 #if defined(__linux__) || defined(__FreeBSD__) || defined(__OpenBSD__) || defined(__GNU__) || defined(__QNX__)
@@ -203,14 +215,13 @@ TEST(EnvVarICDOverrideSetup, XDG) {
     // Set up a layer path that includes default and user-specified locations,
     // so that the test app can find them.  Include some badly specified elements as well.
     // Need to redirect the 'home' directory
-    std::filesystem::path HOME = "/home/fake_home";
-    EnvVarWrapper home_env_var{"HOME", HOME};
-    EnvVarWrapper xdg_config_dirs_env_var{"XDG_CONFIG_DIRS", ":/tmp/goober:::::/tmp/goober/::::"};
-    EnvVarWrapper xdg_config_home_env_var{"XDG_CONFIG_HOME", ":/tmp/goober:::::/tmp/goober2/::::"};
-    EnvVarWrapper xdg_data_dirs_env_var{"XDG_DATA_DIRS", "::::/tmp/goober3:/tmp/goober4/with spaces:::"};
-    EnvVarWrapper xdg_data_home_env_var{"XDG_DATA_HOME", "::::/tmp/goober3:/tmp/goober4/with spaces:::"};
 
-    FrameworkEnvironment env{};
+    FrameworkEnvironment env{FrameworkSettings{}
+                                 .set_home_env_var("/home/fake_home")
+                                 .set_xdg_config_dirs_env_var(":/tmp/goober:::::/tmp/goober/::::")
+                                 .set_xdg_config_home_env_var(":/tmp/goober:::::/tmp/goober2/::::")
+                                 .set_xdg_data_dirs_env_var("::::/tmp/goober3:/tmp/goober4/with spaces:::")
+                                 .set_xdg_data_home_env_var("::::/tmp/goober3:/tmp/goober4/with spaces:::")};
     env.add_icd(TestICDDetails(TEST_ICD_PATH_VERSION_2_EXPORT_ICD_GPDPA)).add_physical_device("physical_device_0");
 
     InstWrapper inst{env.vulkan_functions};
@@ -232,9 +243,7 @@ TEST(EnvVarICDOverrideSetup, XDGContainsJsonFile) {
     // Set up a layer path that includes default and user-specified locations,
     // so that the test app can find them.  Include some badly specified elements as well.
     // Need to redirect the 'home' directory
-    EnvVarWrapper xdg_config_dirs_env_var{"XDG_CONFIG_DIRS", "bad_file.json"};
-
-    FrameworkEnvironment env{};
+    FrameworkEnvironment env{FrameworkSettings{}.set_xdg_config_dirs_env_var("bad_file.json")};
     env.add_icd(TestICDDetails(TEST_ICD_PATH_VERSION_2_EXPORT_ICD_GPDPA)).add_physical_device("physical_device_0");
 
     InstWrapper inst{env.vulkan_functions};
@@ -249,24 +258,27 @@ TEST(EnvVarICDOverrideSetup, TestOnlyAddDriverEnvVar) {
     env.add_icd(TestICDDetails(TEST_ICD_PATH_EXPORT_NONE).set_discovery_type(ManifestDiscoveryType::add_env_var));
     env.get_test_icd(0).physical_devices.emplace_back("pd0");
 
-    InstWrapper inst1{env.vulkan_functions};
-    FillDebugUtilsCreateDetails(inst1.create_info, env.debug_log);
-    inst1.CheckCreate();
+    InstWrapper inst{env.vulkan_functions};
+    FillDebugUtilsCreateDetails(inst.create_info, env.debug_log);
+    inst.CheckCreate();
 
     std::array<VkPhysicalDevice, 1> phys_devs_array;
     uint32_t phys_dev_count = 1;
-    ASSERT_EQ(inst1->vkEnumeratePhysicalDevices(inst1.inst, &phys_dev_count, phys_devs_array.data()), VK_SUCCESS);
+    ASSERT_EQ(inst->vkEnumeratePhysicalDevices(inst.inst, &phys_dev_count, phys_devs_array.data()), VK_SUCCESS);
     ASSERT_EQ(phys_dev_count, 1U);
+}
 
-    env.platform_shim->set_elevated_privilege(true);
+// Test VK_ADD_DRIVER_FILES environment variable with elelvated privileges
+TEST(EnvVarICDOverrideSetup, TestOnlyAddDriverEnvVarRunningWithElevatedPrivileges) {
+    FrameworkEnvironment env{FrameworkSettings{}.set_run_as_if_with_elevated_privleges(true)};
+    env.add_icd(TestICDDetails(TEST_ICD_PATH_EXPORT_NONE).set_discovery_type(ManifestDiscoveryType::add_env_var));
+    env.get_test_icd(0).physical_devices.emplace_back("pd0");
 
-    InstWrapper inst2{env.vulkan_functions};
-    FillDebugUtilsCreateDetails(inst2.create_info, env.debug_log);
-    inst2.CheckCreate(VK_ERROR_INCOMPATIBLE_DRIVER);
+    InstWrapper inst{env.vulkan_functions};
+    FillDebugUtilsCreateDetails(inst.create_info, env.debug_log);
+    inst.CheckCreate(VK_ERROR_INCOMPATIBLE_DRIVER);
 
     EXPECT_TRUE(env.debug_log.find("vkCreateInstance: Found no drivers!"));
-
-    env.platform_shim->set_elevated_privilege(false);
 }
 
 // Test Both VK_DRIVER_FILES and VK_ADD_DRIVER_FILES environment variable
@@ -295,7 +307,7 @@ TEST(EnvVarICDOverrideSetup, TestBothDriverEnvVars) {
 TEST(EnvVarICDOverrideSetup, TestOnlyLayerEnvVar) {
     FrameworkEnvironment env{};
     env.add_icd(TestICDDetails(TEST_ICD_PATH_VERSION_2_EXPORT_ICD_GPDPA)).add_physical_device("physical_device_0");
-    env.platform_shim->redirect_path("/tmp/carol", env.get_folder(ManifestLocation::explicit_layer_env_var).location());
+    env.file_system_manager.add_path_redirect("/tmp/carol", ManifestLocation::explicit_layer_env_var);
 
     const char* layer_name = "TestLayer";
     env.add_explicit_layer(
@@ -312,35 +324,51 @@ TEST(EnvVarICDOverrideSetup, TestOnlyLayerEnvVar) {
     std::string vk_layer_path = ":/tmp/carol::::/:";
     vk_layer_path += (HOME / "/ with spaces/:::::/tandy:");
     EnvVarWrapper layer_path_env_var{"VK_LAYER_PATH", vk_layer_path};
-    InstWrapper inst1{env.vulkan_functions};
-    inst1.create_info.add_layer(layer_name);
-    FillDebugUtilsCreateDetails(inst1.create_info, env.debug_log);
-    inst1.CheckCreate();
+    InstWrapper inst{env.vulkan_functions};
+    inst.create_info.add_layer(layer_name);
+    FillDebugUtilsCreateDetails(inst.create_info, env.debug_log);
+    inst.CheckCreate();
 
     // look for VK_LAYER_PATHS
     EXPECT_TRUE(env.debug_log.find("/tmp/carol"));
     EXPECT_TRUE(env.debug_log.find("/tandy"));
     EXPECT_TRUE(env.debug_log.find((HOME / "/ with spaces/")));
+}
+// Test VK_LAYER_PATH environment variable with elevated privileges
+TEST(EnvVarICDOverrideSetup, TestOnlyLayerEnvVarRunningWithElevatedPrivileges) {
+    FrameworkEnvironment env{FrameworkSettings{}.set_run_as_if_with_elevated_privleges(true)};
+    env.add_icd(TestICDDetails(TEST_ICD_PATH_VERSION_2_EXPORT_ICD_GPDPA)).add_physical_device("physical_device_0");
+    env.file_system_manager.add_path_redirect("/tmp/carol", ManifestLocation::explicit_layer_env_var);
 
-    env.debug_log.clear();
+    const char* layer_name = "TestLayer";
+    env.add_explicit_layer(
+        TestLayerDetails(ManifestLayer{}.add_layer(
+                             ManifestLayer::LayerDescription{}.set_name(layer_name).set_lib_path(TEST_LAYER_PATH_EXPORT_VERSION_2)),
+                         "test_layer.json")
+            .set_discovery_type(ManifestDiscoveryType::env_var));
 
-    env.platform_shim->set_elevated_privilege(true);
+    // Now set up a layer path that includes default and user-specified locations,
+    // so that the test app can find them.  Include some badly specified elements as well.
+    // Need to redirect the 'home' directory
+    std::filesystem::path HOME = "/home/fake_home";
+    EnvVarWrapper home_env_var{"HOME", HOME};
+    std::string vk_layer_path = ":/tmp/carol::::/:";
+    vk_layer_path += (HOME / "/ with spaces/:::::/tandy:");
+    EnvVarWrapper layer_path_env_var{"VK_LAYER_PATH", vk_layer_path};
 
-    InstWrapper inst2{env.vulkan_functions};
-    inst2.create_info.add_layer(layer_name);
-    FillDebugUtilsCreateDetails(inst2.create_info, env.debug_log);
-    inst2.CheckCreate(VK_ERROR_LAYER_NOT_PRESENT);
+    InstWrapper inst{env.vulkan_functions};
+    inst.create_info.add_layer(layer_name);
+    FillDebugUtilsCreateDetails(inst.create_info, env.debug_log);
+    inst.CheckCreate(VK_ERROR_LAYER_NOT_PRESENT);
 
     EXPECT_FALSE(env.debug_log.find("/tmp/carol"));
-
-    env.platform_shim->set_elevated_privilege(false);
 }
 
 // Test VK_ADD_LAYER_PATH environment variable
 TEST(EnvVarICDOverrideSetup, TestOnlyAddLayerEnvVar) {
     FrameworkEnvironment env{};
     env.add_icd(TestICDDetails(TEST_ICD_PATH_VERSION_2_EXPORT_ICD_GPDPA)).add_physical_device("physical_device_0");
-    env.platform_shim->redirect_path("/tmp/carol", env.get_folder(ManifestLocation::explicit_layer_add_env_var).location());
+    env.file_system_manager.add_path_redirect("/tmp/carol", ManifestLocation::explicit_layer_add_env_var);
 
     const char* layer_name = "TestLayer";
     env.add_explicit_layer(
@@ -358,35 +386,52 @@ TEST(EnvVarICDOverrideSetup, TestOnlyAddLayerEnvVar) {
     vk_layer_path += (HOME / "/ with spaces/:::::/tandy:");
     EnvVarWrapper add_layer_path_env_var{"VK_ADD_LAYER_PATH", vk_layer_path};
 
-    InstWrapper inst1{env.vulkan_functions};
-    inst1.create_info.add_layer(layer_name);
-    FillDebugUtilsCreateDetails(inst1.create_info, env.debug_log);
-    inst1.CheckCreate();
+    InstWrapper inst{env.vulkan_functions};
+    inst.create_info.add_layer(layer_name);
+    FillDebugUtilsCreateDetails(inst.create_info, env.debug_log);
+    inst.CheckCreate();
 
     // look for VK_ADD_LAYER_PATHS
     EXPECT_TRUE(env.debug_log.find("/tmp/carol"));
     EXPECT_TRUE(env.debug_log.find("/tandy"));
     EXPECT_TRUE(env.debug_log.find((HOME / "/ with spaces/")));
+}
 
-    env.debug_log.clear();
+// Test VK_ADD_LAYER_PATH environment variable with elevated privileges
+TEST(EnvVarICDOverrideSetup, TestOnlyAddLayerEnvVarRunningWithElevatedPrivileges) {
+    FrameworkEnvironment env{FrameworkSettings{}.set_run_as_if_with_elevated_privleges(true)};
+    env.add_icd(TestICDDetails(TEST_ICD_PATH_VERSION_2_EXPORT_ICD_GPDPA)).add_physical_device("physical_device_0");
+    env.file_system_manager.add_path_redirect("/tmp/carol", ManifestLocation::explicit_layer_add_env_var);
 
-    env.platform_shim->set_elevated_privilege(true);
+    const char* layer_name = "TestLayer";
+    env.add_explicit_layer(
+        TestLayerDetails(ManifestLayer{}.add_layer(
+                             ManifestLayer::LayerDescription{}.set_name(layer_name).set_lib_path(TEST_LAYER_PATH_EXPORT_VERSION_2)),
+                         "test_layer.json")
+            .set_discovery_type(ManifestDiscoveryType::add_env_var));
 
-    InstWrapper inst2{env.vulkan_functions};
-    inst2.create_info.add_layer(layer_name);
-    FillDebugUtilsCreateDetails(inst2.create_info, env.debug_log);
-    inst2.CheckCreate(VK_ERROR_LAYER_NOT_PRESENT);
+    // Set up a layer path that includes default and user-specified locations,
+    // so that the test app can find them.  Include some badly specified elements as well.
+    // Need to redirect the 'home' directory
+    std::filesystem::path HOME = "/home/fake_home";
+    EnvVarWrapper home_env_var{"HOME", HOME};
+    std::string vk_layer_path = ":/tmp/carol::::/:";
+    vk_layer_path += (HOME / "/ with spaces/:::::/tandy:");
+    EnvVarWrapper add_layer_path_env_var{"VK_ADD_LAYER_PATH", vk_layer_path};
+
+    InstWrapper inst{env.vulkan_functions};
+    inst.create_info.add_layer(layer_name);
+    FillDebugUtilsCreateDetails(inst.create_info, env.debug_log);
+    inst.CheckCreate(VK_ERROR_LAYER_NOT_PRESENT);
 
     EXPECT_FALSE(env.debug_log.find("/tmp/carol"));
-
-    env.platform_shim->set_elevated_privilege(false);
 }
 
 // Test VK_IMPLICIT_LAYER_PATH environment variable
 TEST(EnvVarICDOverrideSetup, TestOnlyImplicitLayerEnvVar) {
     FrameworkEnvironment env{};
     env.add_icd(TestICDDetails(TEST_ICD_PATH_VERSION_2_EXPORT_ICD_GPDPA)).add_physical_device("physical_device_0");
-    env.platform_shim->redirect_path("/tmp/carol", env.get_folder(ManifestLocation::implicit_layer_env_var).location());
+    env.file_system_manager.add_path_redirect("/tmp/carol", ManifestLocation::implicit_layer_env_var);
 
     const char* layer_name = "TestLayer";
     env.add_implicit_layer(TestLayerDetails(ManifestLayer{}.add_layer(ManifestLayer::LayerDescription{}
@@ -404,39 +449,57 @@ TEST(EnvVarICDOverrideSetup, TestOnlyImplicitLayerEnvVar) {
     std::string vk_implicit_layer_path = ":/tmp/carol::::/:";
     vk_implicit_layer_path += (HOME / "/ with spaces/:::::/tandy:");
     EnvVarWrapper implicit_layer_path_env_var{"VK_IMPLICIT_LAYER_PATH", vk_implicit_layer_path};
-    InstWrapper inst1{env.vulkan_functions};
-    FillDebugUtilsCreateDetails(inst1.create_info, env.debug_log);
-    inst1.CheckCreate();
+    InstWrapper inst{env.vulkan_functions};
+    FillDebugUtilsCreateDetails(inst.create_info, env.debug_log);
+    inst.CheckCreate();
 
-    auto active_layers1 = inst1.GetActiveLayers(inst1.GetPhysDev(), 1);
-    ASSERT_TRUE(string_eq(active_layers1.at(0).layerName, layer_name));
+    auto active_layers = inst.GetActiveLayers(inst.GetPhysDev(), 1);
+    ASSERT_TRUE(string_eq(active_layers.at(0).layerName, layer_name));
 
     // look for VK_IMPLICIT_LAYER_PATHS
     EXPECT_TRUE(env.debug_log.find("/tmp/carol"));
     EXPECT_TRUE(env.debug_log.find("/tandy"));
     EXPECT_TRUE(env.debug_log.find((HOME / "/ with spaces/")));
+}
 
-    env.debug_log.clear();
+// Test VK_IMPLICIT_LAYER_PATH environment variable run with elevated privileges
+TEST(EnvVarICDOverrideSetup, TestOnlyImplicitLayerEnvVarRunningWithElevatedPrivileges) {
+    FrameworkEnvironment env{FrameworkSettings{}.set_run_as_if_with_elevated_privleges(true)};
+    env.add_icd(TestICDDetails(TEST_ICD_PATH_VERSION_2_EXPORT_ICD_GPDPA)).add_physical_device("physical_device_0");
+    env.file_system_manager.add_path_redirect("/tmp/carol", ManifestLocation::implicit_layer_env_var);
 
-    env.platform_shim->set_elevated_privilege(true);
+    const char* layer_name = "TestLayer";
+    env.add_implicit_layer(TestLayerDetails(ManifestLayer{}.add_layer(ManifestLayer::LayerDescription{}
+                                                                          .set_name(layer_name)
+                                                                          .set_lib_path(TEST_LAYER_PATH_EXPORT_VERSION_2)
+                                                                          .set_disable_environment("DISABLE_ENV")),
+                                            "test_layer.json")
+                               .set_discovery_type(ManifestDiscoveryType::env_var));
 
-    InstWrapper inst2{env.vulkan_functions};
-    FillDebugUtilsCreateDetails(inst2.create_info, env.debug_log);
-    inst2.CheckCreate();
+    // Now set up a layer path that includes default and user-specified locations,
+    // so that the test app can find them.  Include some badly specified elements as well.
+    // Need to redirect the 'home' directory
+    std::filesystem::path HOME = "/home/fake_home";
+    EnvVarWrapper home_env_var{"HOME", HOME};
+    std::string vk_implicit_layer_path = ":/tmp/carol::::/:";
+    vk_implicit_layer_path += (HOME / "/ with spaces/:::::/tandy:");
+    EnvVarWrapper implicit_layer_path_env_var{"VK_IMPLICIT_LAYER_PATH", vk_implicit_layer_path};
 
-    auto active_layers2 = inst2.GetActiveLayers(inst2.GetPhysDev(), 0);
-    ASSERT_TRUE(active_layers2.empty());
+    InstWrapper inst{env.vulkan_functions};
+    FillDebugUtilsCreateDetails(inst.create_info, env.debug_log);
+    inst.CheckCreate();
+
+    auto active_layers = inst.GetActiveLayers(inst.GetPhysDev(), 0);
+    ASSERT_TRUE(active_layers.empty());
 
     EXPECT_FALSE(env.debug_log.find("/tmp/carol"));
-
-    env.platform_shim->set_elevated_privilege(false);
 }
 
 // Test VK_ADD_IMPLICIT_LAYER_PATH environment variable
 TEST(EnvVarICDOverrideSetup, TestOnlyAddImplicitLayerEnvVar) {
     FrameworkEnvironment env{};
     env.add_icd(TestICDDetails(TEST_ICD_PATH_VERSION_2_EXPORT_ICD_GPDPA)).add_physical_device("physical_device_0");
-    env.platform_shim->redirect_path("/tmp/carol", env.get_folder(ManifestLocation::implicit_layer_add_env_var).location());
+    env.file_system_manager.add_path_redirect("/tmp/carol", ManifestLocation::implicit_layer_add_env_var);
 
     const char* layer_name = "TestLayer";
     env.add_implicit_layer(TestLayerDetails(ManifestLayer{}.add_layer(ManifestLayer::LayerDescription{}
@@ -453,7 +516,7 @@ TEST(EnvVarICDOverrideSetup, TestOnlyAddImplicitLayerEnvVar) {
     EnvVarWrapper home_env_var{"HOME", HOME};
     std::string vk_add_implicit_layer_path = ":/tmp/carol::::/:";
     vk_add_implicit_layer_path += (HOME / "/ with spaces/:::::/tandy:");
-    EnvVarWrapper add_implicit_layer_path_env_var{"VK_ADD_LAYER_PATH", vk_add_implicit_layer_path};
+    EnvVarWrapper add_implicit_layer_path_env_var{"VK_ADD_IMPLICIT_LAYER_PATH", vk_add_implicit_layer_path};
 
     InstWrapper inst1{env.vulkan_functions};
     FillDebugUtilsCreateDetails(inst1.create_info, env.debug_log);
@@ -466,21 +529,39 @@ TEST(EnvVarICDOverrideSetup, TestOnlyAddImplicitLayerEnvVar) {
     EXPECT_TRUE(env.debug_log.find("/tmp/carol"));
     EXPECT_TRUE(env.debug_log.find("/tandy"));
     EXPECT_TRUE(env.debug_log.find((HOME / "/ with spaces/")));
+}
 
-    env.debug_log.clear();
+// Test VK_ADD_IMPLICIT_LAYER_PATH environment variable running with elevated privileges
+TEST(EnvVarICDOverrideSetup, TestOnlyAddImplicitLayerEnvVarWithElevatedPrivileges) {
+    FrameworkEnvironment env{FrameworkSettings{}.set_run_as_if_with_elevated_privleges(true)};
+    env.add_icd(TestICDDetails(TEST_ICD_PATH_VERSION_2_EXPORT_ICD_GPDPA)).add_physical_device("physical_device_0");
+    env.file_system_manager.add_path_redirect("/tmp/carol", ManifestLocation::implicit_layer_add_env_var);
 
-    env.platform_shim->set_elevated_privilege(true);
+    const char* layer_name = "TestLayer";
+    env.add_implicit_layer(TestLayerDetails(ManifestLayer{}.add_layer(ManifestLayer::LayerDescription{}
+                                                                          .set_name(layer_name)
+                                                                          .set_lib_path(TEST_LAYER_PATH_EXPORT_VERSION_2)
+                                                                          .set_disable_environment("DISABLE_ENV")),
+                                            "test_layer.json")
+                               .set_discovery_type(ManifestDiscoveryType::add_env_var));
 
-    InstWrapper inst2{env.vulkan_functions};
-    FillDebugUtilsCreateDetails(inst2.create_info, env.debug_log);
-    inst2.CheckCreate();
+    // Set up a layer path that includes default and user-specified locations,
+    // so that the test app can find them.  Include some badly specified elements as well.
+    // Need to redirect the 'home' directory
+    std::filesystem::path HOME = "/home/fake_home";
+    EnvVarWrapper home_env_var{"HOME", HOME};
+    std::string vk_add_implicit_layer_path = ":/tmp/carol::::/:";
+    vk_add_implicit_layer_path += (HOME / "/ with spaces/:::::/tandy:");
+    EnvVarWrapper add_implicit_layer_path_env_var{"VK_ADD_IMPLICIT_LAYER_PATH", vk_add_implicit_layer_path};
 
-    auto active_layers2 = inst2.GetActiveLayers(inst2.GetPhysDev(), 0);
-    ASSERT_TRUE(active_layers2.empty());
+    InstWrapper inst{env.vulkan_functions};
+    FillDebugUtilsCreateDetails(inst.create_info, env.debug_log);
+    inst.CheckCreate();
+
+    auto active_layers = inst.GetActiveLayers(inst.GetPhysDev(), 0);
+    ASSERT_TRUE(active_layers.empty());
 
     EXPECT_FALSE(env.debug_log.find("/tmp/carol"));
-
-    env.platform_shim->set_elevated_privilege(false);
 }
 
 #endif
